refactor: UDecalComponent rendering method signature
Updated the `RenderOnActor` method to accept a `FViewport*` parameter instead of an `AActor*`. Added a check for viewport show flags to conditionally render static meshes. Adjusted the call to `RenderOnActor` in `UWorld::RenderViewports` to reflect the new method signature. The destructor for `UDecalComponent` is now defined as an empty function.